back-end/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tortoise-orm==0.20.0
aerich==0.7.2
python-dotenv==1.0.0
pydantic>=2.0.0
